Actions: Break latch on modifier/group changes
Right now, only a few actions result in breaking a key latch. Add
modifier/group actions to this, as presumably the effect of the latch
should not last beyond already shifting the modifier/group state.

This was discovered with a real-world use in an input device for the
physically impaired, where shift latches, but pressing shift twice goes
to caps lock (note shift vs. caps lock, so not pure latch-to-lock). This
is possible to implement with current keymap files, but locking caps
lock would not break the shift latch. Hence, the following key sequence:

[Shift] [a -> a] [Shift] [Shift] [b -> b] [c -> C]

would occur, as the modifier state when pressing b was lock locked and
shift latched. With this patch, the sequence is correct:

[Shift] [a -> a] [Shift] [Shift -> Lock] [b -> B] [c -> C]

Signed-off-by: Daniel Stone <daniels@collabora.com>