test: untangle print-compiled-keymap from the test headers
Commit 16c84cdd819db516fff089c76b99248fb7dd4e8c removed the getopt handling for
RMLVO arguments, so now this tool only takes a keymap file and compiles it.

Using test helpers to init the context gives it fairly specific behavior; unless
the user sets the right environment variables and/or calls it from the right
PWD, it may or may not include the test data.

Let's drop this behavior, make it a default tool to compile a keymap. If there
is a specific need to modify the include paths, we can add this later.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>