tools: drop getopt vs getopt_long differentiation
On all platforms we build on where getopt.h is available, getopt_long is also
available. Only Windows doesn't have either but that's no reason for us to
differentiate between the two.

If we need to special-case getopt vs getopt_long, it's probably better to
implement our own cross-platform version of it and use that.

Fixes #161

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>