state: support querying whether virtual modifiers are active
Previously it was not possible to query the status of virtual
modifiers with the following functions:
- xkb_state_mod_index_is_active
- xkb_state_mod_indices_are_active
- xkb_state_mod_name_is_active
- xkb_state_mod_names_are_active

Note that it may overmatch if some modifier mappings overlap.
For example, the default “us” PC layout maps “Alt” and “Meta”
to the real modifier “Mod1”; thus “Mod1”, “Alt” and “Meta”
modifiers will return the same result with these functions.