Interactive tools: Escape control character for Unicode output
Currently the interactive tools print the string result of key strokes
as it is, without any escape. This is especially annoying for trivial
keysyms such as: Return, BackSpace and Escape.

Fix this by displaying the Unicode code point notation (e.g U+000D for
Return) for single control characters from the C0 set and DEL.

This is a hack: ideally we would like to escape any non-printable
character in the utf-8 string.