Add new function xkb_keysym_with_format_from_name
`xkb_keysym_from_name` does not provide the successful type of
parsing, i.e. if a name, a Unicode codepoint or an hexadecimal
number.

A the new function `xkb_keysym_with_format_from_name`, which acts as
`xkb_keysym_from_name`, but it also provide the type of successful
parsing.

Also improve the Unicode parsing tests.