rules: Add support for :all qualifier
Some layout options require to be applied to every group to maintain
consistency (e.g. a group switcher). Currently this must be done manually
for all layout indexes. This is error prone and prevents the increase of
the maximum group count.

This commit introduces the :all qualifier for kccgst values. When a rule
with this qualifier applies, it will expands for every layout, e.g.
`+group(toggle):all` would expand to `+group(toggle):1+group(toggle):2`
if there are 2 layouts, etc.